refactor: code quality improvements across utils modules
- Add type annotations to resolve_config and _flatten_lists
- Remove dead code in get_argument_value and unused version param
- Refactor handle_snapshot to accept SnapshotEvaluation directly
- Extract output path resolution to utils/paths.py
- Remove legacy_cli_output_flags from RunConfig (computed at use site)
- Replace print() with typer.echo() in toml error messages
- Import code_complexity from _complexipy directly
- Add from __future__ import annotations to all utils files
- Narrow _is_legacy_snapshot type from List[Any] to List[FileComplexity]
- Fix ty check error with toml_config None narrowing
